Color-scheme store fix.
