+++ src/lib/stores.ts
@@ -21,14 +21,4 @@ export const errors: Writable<Error[]> = writable([]);
  * Represents the color-scheme store.
  * Stores the color-scheme value.
  */
-const userTheme = browser && localStorage.getItem('color-scheme');
-export const theme = writable(userTheme ?? 'dark');
-
-export function toggleTheme() {
-	theme.update((currentTheme) => {
-		const newTheme = currentTheme === 'dark' ? 'light' : 'dark';
-		document.documentElement.setAttribute('color-scheme', newTheme);
-		localStorage.setItem('color-scheme', newTheme);
-		return newTheme;
-	});
-}
+export const theme = writable((browser && localStorage.getItem('color-scheme')) ?? 'light');
